Adds connect with Stripe to admin billing and removes stripe connect test component and route from tenant
